Provide an upgrade path for the module merge that happen in 4.0.0
