+++ vertx-auth-common/src/main/asciidoc/dataobjects.adoc
@@ -14,6 +14,39 @@
 ^|Name | Type ^| Description
 |===
 
+[[JWTOptions]]
+== JWTOptions
+
+++++
+ Options related to creation of new tokens.
+
+ If any expiresInMinutes, audience, subject, issuer are not provided, there is no default.
+ The jwt generated won't include those properties in the payload.
+
+ Generated JWTs will include an iat claim by default unless noTimestamp is specified.
+++++
+'''
+
+[cols=">25%,25%,50%"]
+[frame="topbot"]
+|===
+^|Name | Type ^| Description
+|[[algorithm]]`@algorithm`|`String`|-
+|[[audience]]`@audience`|`Array of String`|-
+|[[audiences]]`@audiences`|`Array of String`|-
+|[[expiresInMinutes]]`@expiresInMinutes`|`Number (int)`|-
+|[[expiresInSeconds]]`@expiresInSeconds`|`Number (int)`|-
+|[[header]]`@header`|`Json object`|-
+|[[ignoreExpiration]]`@ignoreExpiration`|`Boolean`|-
+|[[issuer]]`@issuer`|`String`|-
+|[[leeway]]`@leeway`|`Number (int)`|-
+|[[noTimestamp]]`@noTimestamp`|`Boolean`|-
+|[[permissions]]`@permissions`|`Array of String`|+++
+The permissions of this token.
++++
+|[[subject]]`@subject`|`String`|-
+|===
+
 [[KeyStoreOptions]]
 == KeyStoreOptions
