+++ vertx-auth-jwt/src/main/java/io/vertx/ext/auth/jwt/JWTAuth.java
@@ -21,8 +21,8 @@ import io.vertx.codegen.annotations.VertxGen;
 import io.vertx.core.Vertx;
 import io.vertx.core.json.JsonObject;
 import io.vertx.ext.auth.AuthProvider;
+import io.vertx.ext.auth.JWTOptions;
 import io.vertx.ext.auth.jwt.impl.JWTAuthProviderImpl;
-import io.vertx.ext.jwt.JWTOptions;
 
 /**
  * Factory interface for creating JWT based {@link io.vertx.ext.auth.AuthProvider} instances.
