+++ vertx-auth-jwt/src/main/java/io/vertx/ext/auth/jwt/JWTAuthOptions.java
@@ -20,7 +20,7 @@ import io.vertx.core.json.JsonObject;
 import io.vertx.ext.auth.KeyStoreOptions;
 import io.vertx.ext.auth.PubSecKeyOptions;
 import io.vertx.ext.auth.SecretOptions;
-import io.vertx.ext.jwt.JWTOptions;
+import io.vertx.ext.auth.JWTOptions;
 
 import java.util.ArrayList;
 import java.util.List;
