+++ vertx-jwt/pom.xml
@@ -31,6 +31,13 @@
     <doc.skip>true</doc.skip>
   </properties>
 
+  <dependencies>
+    <dependency>
+      <groupId>io.vertx</groupId>
+      <artifactId>vertx-auth-common</artifactId>
+    </dependency>
+  </dependencies>
+
   <build>
     <plugins>
       <plugin>
